requirements.txt
streamlit
ultralytics
Pillow
pathlib
pandas